Base-template/requirements.txt
asgiref==3.3.1
Django==3.1.5
django-crispy-forms==1.10.0
django-utils-six==2.0
mysql-connector==2.2.9
Pillow==8.1.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
